refactor hooks.
